make SslClientAuthenticationOptionsExtensions internal
